Update 14.biological-interpretation.md
I have finished with the topics I wanted to contribute with. Next for me is (I) adding references (II) maybe producing another figure in the enrichment part. Does the table need to be in a figure? Or is there a way to produce tables?

It will be done by next week or sooner. 

Best,