Update 15.biological-interpretation.md (#67)
* Update 15.biological-interpretation.md

Added a section on methods for orthogonal validation experiments. Since this is a vast topic, the section is not comprehensive and still lacks a bunch of references. Moreover, some of th content might go well with the 'Experiment Types' section of the review?

* Update 15.biological-interpretation.md

added some references and adjusted wording

* Update 15.biological-interpretation.md

just refining the text, adding references etc. Still work in progress.

* Update 15.biological-interpretation.md

I am done with the first draft. Far from perfect, but happy to add to it and amend.

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Update content/15.biological-interpretation.md

* Apply suggestions from code review

---------

Co-authored-by: jessegmeyerlab <65188012+jessegmeyerlab@users.noreply.github.com>